+++ builder/style.css
@@ -10,6 +10,7 @@
 
 @import url('~@jupyterlab/codemirror-extension/style/index.css');
 @import url('~@jupyterlab/docmanager-extension/style/index.css');
+@import url('~@jupyterlab/fileeditor-extension/style/index.css');
 @import url('~@jupyterlab/mainmenu-extension/style/index.css');
 @import url('~@jupyterlab/notebook-extension/style/index.css');
 @import url('~@jupyterlab/rendermime-extension/style/index.css');
