+++ packages/application-extension/style/base.css
@@ -8,3 +8,7 @@
   flex-grow: 1;
   flex-shrink: 1;
 }
+
+.jp-Document {
+  height: 100%;
+}
