+++ packages/docmanager-extension/src/index.ts
@@ -6,7 +6,7 @@ import {
   JupyterFrontEndPlugin
 } from '@jupyterlab/application';
 
-import { PageConfig } from '@jupyterlab/coreutils';
+import { PageConfig, PathExt } from '@jupyterlab/coreutils';
 
 import { IDocumentManager } from '@jupyterlab/docmanager';
 
@@ -39,7 +39,9 @@ const opener: JupyterFrontEndPlugin<void> = {
         docOpen.call(docManager, path, widgetName, kernel, options);
         return;
       }
-      window.open(`${baseUrl}classic/notebooks/${path}`);
+      const ext = PathExt.extname(path);
+      const route = ext === '.ipynb' ? 'notebooks' : 'edit';
+      window.open(`${baseUrl}classic/${route}/${path}`);
       return undefined;
     };
   }
